% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adults}
\alias{adults}
\title{Adult trees}
\format{
\subsection{\code{adults}}{

A data frame with 59 rows and 35 columns:
\describe{
\item{Sample}{Pie215}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
adults
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
