\name{calculateDeletionlengths}
\alias{calculateDeletionlengths}
\title{Calculates the length of the deletions in the DNV object}
\usage{
calculateDeletionlengths(DNVobject)
}
\description{
This function will automatically grab only the deletions from the DNV object for the calculation of the length of the deletions.
}
\examples{
calculateDeletionlengths(DNVobject)
}
\value{
Returns the length of the deletions, in the form of an object, observed in the DNV object. It also returns a barplot of the deletion lengths.
}
