% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mock_dataset.R
\name{create_mock_parents}
\alias{create_mock_parents}
\title{Title}
\usage{
create_mock_parents(nmarkers = 10, ntrees = 100, nvariants = 4, maf = NULL)
}
\arguments{
\item{nmarkers}{number of markers}

\item{ntrees}{number of trees}

\item{nvariants}{number of trees}

\item{maf}{minimum allele frequency}
}
\value{
a list
}
\description{
Title
}
