% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mock_dataset.R
\name{create_mock_progeny}
\alias{create_mock_progeny}
\title{Title}
\usage{
create_mock_progeny(info, fparents, mparents, prog)
}
\arguments{
\item{info}{mock parents, as generated from \code{create_mock_parents} function}

\item{fparents}{number of female parents}

\item{mparents}{number of male parents}

\item{prog}{number of progeny??}
}
\value{
a data table
}
\description{
Title
}
