% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_duplicates.R
\name{exclude_duplicates}
\alias{exclude_duplicates}
\title{Report duplicates}
\usage{
exclude_duplicates(parents, adults = NULL, progeny = NULL)
}
\arguments{
\item{parents}{a data.frame}

\item{adults}{a data.frame}

\item{progeny}{a data.frame}
}
\value{
a data.frame
}
\description{
Report duplicates
}
