\name{extractMNVs}
\alias{extractMNVs}
\title{Extract MNVs from DNV object}
\usage{
extractMNVs(DNVobject = NULL)
}
\arguments{
\item{DNVobject}{a DNVobject}
}
\description{
Extracts multi-nucleotide variants (MNVs) out from a DNV object generated using the readDNV function.
}
\examples{
extractMNVs(DNVobject = NULL)
}
\value{
Returns a DNV object containing only MNVs
}
