\name{extractSNVs}
\alias{extractSNVs}
\title{Extract SNVs from DNV object}
\usage{
extractSNVs(DNVobject)
}
\arguments{
\item{DNVobject}{a DNVobject}
}
\description{
Extracts single-nucleotide variants (SNVs) out from a DNV object generated using the readDNV function.
}
\examples{
extractSNVs(DNVobject)
}
\value{
Returns a DNV object containing only SNVs
}
