% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parents.R
\name{find_parents}
\alias{find_parents}
\title{Identify relationships between parents and progenies}
\usage{
find_parents(adults, progeny)
}
\arguments{
\item{adults}{a data.frame containing}

\item{progeny}{a data.frame}
}
\value{
a data.frame
}
\description{
Identify relationships between parents and progenies
}
