% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_duplicates.R
\name{identify_duplicates}
\alias{identify_duplicates}
\title{Idntify duplicates in genotypes (i.e. parents or progenies)}
\usage{
identify_duplicates(genotypes, abbr = NULL)
}
\arguments{
\item{genotypes}{a data.frame with the genotypes}

\item{abbr}{a string with abbreviation to use}
}
\value{
a data.frame
}
\description{
Idntify duplicates in genotypes (i.e. parents or progenies)
}
