% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parents}
\alias{parents}
\title{Tree parents data set}
\format{
\subsection{\code{parents}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
parents
}
\description{
An example of tree parents data set
}
\keyword{datasets}
