\name{extractIndividual}
\alias{extractIndividual}
\title{Extract individual, by id, from DNV object}
\usage{
extractIndividual(DNVobject)
}
\description{
Extracts the DNVs out of a dnvObject from a particular individual.
}
\examples{
extractIndividual(DNVobject)
}
\value{
Returns a DNV object containing only DNVs in the specified individual.
}
